+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased 
+## Unreleased
 
-### Added 
+### Added
 
+- Added the `rename` command which renames contracts
 - Added the `pack` command which builds and prepares a publish-ready package of contracts' wrappers
 
 ## [0.33.1] - 2025-05-16